Organize properly the `Tag` data models
