+++ shared/data/model/src/main/kotlin/com/ivy/data/model/Tag.kt
@@ -1,11 +1,14 @@
 package com.ivy.data.model
 
+import androidx.compose.runtime.Immutable
+import com.ivy.data.model.primitive.AssociationId
 import com.ivy.data.model.primitive.ColorInt
 import com.ivy.data.model.primitive.IconAsset
 import com.ivy.data.model.primitive.NotBlankTrimmedString
-import com.ivy.data.model.primitive.TagId
 import com.ivy.data.model.sync.Syncable
+import com.ivy.data.model.sync.UniqueId
 import java.time.Instant
+import java.util.UUID
 
 data class Tag(
     override val id: TagId,
@@ -18,3 +21,16 @@ data class Tag(
     override val lastUpdated: Instant,
     override val removed: Boolean
 ) : Syncable, Reorderable
+
+
+@Suppress("DataClassTypedIDs")
+data class TagAssociation(
+    override val id: TagId,
+    val associatedId: AssociationId,
+    override val lastUpdated: Instant,
+    override val removed: Boolean,
+) : Syncable
+
+@JvmInline
+@Immutable
+value class TagId(override val value: UUID) : UniqueId
