--- shared/data/model/src/main/kotlin/com/ivy/data/model/TagAssociation.kt
@@ -1,14 +1,0 @@
-package com.ivy.data.model
-
-import com.ivy.data.model.primitive.AssociationId
-import com.ivy.data.model.primitive.TagId
-import com.ivy.data.model.sync.Syncable
-import java.time.Instant
-
-@Suppress("DataClassTypedIDs")
-data class TagAssociation(
-    override val id: TagId,
-    val associatedId: AssociationId,
-    override val lastUpdated: Instant,
-    override val removed: Boolean,
-) : Syncable
