+++ shared/data/model/src/main/kotlin/com/ivy/data/model/Transaction.kt
@@ -1,7 +1,6 @@
 package com.ivy.data.model
 
 import com.ivy.data.model.primitive.NotBlankTrimmedString
-import com.ivy.data.model.primitive.TagId
 import com.ivy.data.model.sync.Syncable
 import com.ivy.data.model.sync.UniqueId
 import java.time.Instant
