+++ temp/legacy-code/src/main/java/com/ivy/legacy/ui/component/tags/AddOrEditTagModal.kt
@@ -29,7 +29,7 @@ import androidx.compose.ui.unit.dp
 import arrow.core.raise.either
 import com.ivy.data.model.Tag
 import com.ivy.data.model.primitive.NotBlankTrimmedString
-import com.ivy.data.model.primitive.TagId
+import com.ivy.data.model.TagId
 import com.ivy.wallet.ui.theme.components.DeleteButton
 import com.ivy.wallet.ui.theme.components.IvyTitleTextField
 import com.ivy.wallet.ui.theme.modal.IvyModal
